refactor: Split interactions.rs into directory module structure
Refactors the large genai-client/src/models/interactions.rs (3664 lines) into
a directory module with logical file separation for improved maintainability.

New structure:
- mod.rs (~27 lines) - Re-exports all 18 public types
- content.rs (~550 lines) - InteractionContent enum + custom serde impls
- request.rs (~86 lines) - CreateInteractionRequest, GenerationConfig, InteractionInput
- response.rs (~700 lines) - InteractionResponse + 27 helper methods
- metadata.rs (~120 lines) - Grounding and URL metadata types
- streaming.rs (~48 lines) - StreamChunk and SSE event types
- tests.rs (~1660 lines) - All 66 unit tests

The public API is unchanged - all types remain accessible at the same import paths.
