+++ genai-client/src/models/interactions/request.rs
@@ -1,0 +1,85 @@
+//! Request types for creating interactions.
+
+use serde::{Deserialize, Serialize};
+
+use super::content::InteractionContent;
+use crate::models::shared::Tool;
+
+/// Input for an interaction - can be a simple string or array of content
+#[derive(Clone, Serialize, Deserialize, Debug)]
+#[serde(untagged)]
+pub enum InteractionInput {
+    /// Simple text input
+    Text(String),
+    /// Array of content objects
+    Content(Vec<InteractionContent>),
+}
+
+/// Generation configuration for model behavior
+#[derive(Clone, Serialize, Deserialize, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct GenerationConfig {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub temperature: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub max_output_tokens: Option<i32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub top_p: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub top_k: Option<i32>,
+    /// Thinking level: "minimal", "low", "medium", "high"
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub thinking_level: Option<String>,
+}
+
+/// Request body for the Interactions API endpoint
+#[derive(Clone, Serialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CreateInteractionRequest {
+    /// Model name (e.g., "gemini-3-flash-preview") - mutually exclusive with agent
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub model: Option<String>,
+
+    /// Agent name (e.g., "deep-research-pro-preview-12-2025") - mutually exclusive with model
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub agent: Option<String>,
+
+    /// The input for this interaction
+    pub input: InteractionInput,
+
+    /// Reference to a previous interaction for stateful conversations
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub previous_interaction_id: Option<String>,
+
+    /// Tools available for function calling
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub tools: Option<Vec<Tool>>,
+
+    /// Response modalities (e.g., ["IMAGE"])
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub response_modalities: Option<Vec<String>>,
+
+    /// JSON schema for structured output
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub response_format: Option<serde_json::Value>,
+
+    /// Model configuration
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub generation_config: Option<GenerationConfig>,
+
+    /// Enable streaming responses
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub stream: Option<bool>,
+
+    /// Background execution mode (agents only)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub background: Option<bool>,
+
+    /// Persist interaction data (default: true)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub store: Option<bool>,
+
+    /// System instruction for the model
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub system_instruction: Option<InteractionInput>,
+}
