+++ genai-client/src/models/interactions/streaming.rs
@@ -1,0 +1,47 @@
+//! Streaming types for SSE responses.
+
+use serde::Deserialize;
+
+use super::content::InteractionContent;
+use super::response::{InteractionResponse, InteractionStatus};
+
+/// A chunk from the streaming API
+///
+/// During streaming, the API sends different types of events:
+/// - `Delta`: Incremental content updates (text, thought, function_call, etc.)
+/// - `Complete`: The final complete interaction response
+#[derive(Clone, Debug)]
+pub enum StreamChunk {
+    /// Incremental content update
+    Delta(InteractionContent),
+    /// Complete interaction response (final event)
+    Complete(InteractionResponse),
+}
+
+/// Wrapper for SSE streaming events from the Interactions API
+///
+/// The API returns different event types during streaming:
+/// - `content.delta`: Contains incremental content in the `delta` field
+/// - `interaction.complete`: Contains the full interaction in the `interaction` field
+#[derive(Clone, Deserialize, Debug)]
+#[serde(rename_all = "snake_case")]
+pub struct InteractionStreamEvent {
+    /// Event type (e.g., "content.delta", "interaction.complete")
+    pub event_type: String,
+
+    /// The full interaction data (present in "interaction.complete" events)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub interaction: Option<InteractionResponse>,
+
+    /// Incremental content delta (present in "content.delta" events)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub delta: Option<InteractionContent>,
+
+    /// Interaction ID (present in various events)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub interaction_id: Option<String>,
+
+    /// Status (present in status update events)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub status: Option<InteractionStatus>,
+}
